This massive patch solves most of the issues that prevented the client from working on bronco and paradise servers simultaneously.  The end goal is to remove all the #ifdef PARADISE defines and have a unified client.  These patch notes will describe some of the obstacles, and what was done to work around them.
Obstacles/Solutions:
1) One obstacle was the reliance on MAX* values in the client for things like
MAXPLAYER, MAXTORP, and MAXPHASER.  The solution was to copy the paradise client and define both a theoretical server maximum (i.e. MAXPLAYER) and then the actual value defined by the server (nplayers).  Thus, many checks in the code were changed to use these new dynamic variables.  MAXPLAYER is now 257 instead of 36.  But the default value for nplayers is 36, and will only be changed if the client receives a paradise game parameters packet.  So there is no impact on playing on bronco servers since the default value is always the same.  Likewise, ntorps, nplasmas and nphasers are now dynamic, with the default values being the bronco defaults.
2) Another obstacle was conflicting defines for some server packets.   Namely
SP_GENERIC_32/SP_MOTD_PIC and SP_FLAGS_ALL/SP_STATS2.  The solution was to treat these packets as having dynamic size, with the size being determined on whether the server is paradise or not.  One complication was that a paradise server will send the client a SP_STATS2 packet before it even acknowledges the client is paradise comptable, thus corrupting the data stream!  The solution is to treat all SP_FLAGS_ALL/SP_STATS2 packets as being SP_STATS2 if the login packet was not received yet (the login packet identifies whether the server is paradise or not).
3) Yet another obstacle was conflicting player flag states, namely PFOBSERV/PFWARP, and PFTWARP/PFSNAKE.  The solution was to change all checks in the code involving observers and twarp and paradise warp to check if the server is paradise or not.
4) A minor obstacle was overlapping p_whydead states between bronco and paradise.  Solved by a simple if (paradise) check.

General Changes:
1) Most #ifdef PARADISE removed and replaced, where necessary, by if (paradise)
checks.  Some areas which required more work were the planet window, rank window, motd paging and layout, initializing and sorting the playerlist.
2) Load all paradise bitmaps upon startup.
3) Struct ship updated to include fields the server has been sending via ship cap
packets but we have been ignoring.  Will be necessary for future changes to how
ship caps are handled.
4) Initialization of most data structs (players, torps, phasers, plasmas, teams)
now moved into their own functions rather than all being done in openmem().  In
paradise sometimes they need to be reinitialized, so this was necessary.
5) New funcion isMe() to check if a player X is me, or if I am an observer locked
onto player X.
6) Metaserver now lists paradise servers.
7) Added afterburner/warpprep/impulse indicator to the plain text dashboard.
8) Added a bunch of safety checks for many of the packet types.
9) Added a bunch of the new packet types to the print_packet function.
10) Removed unused compute_extension1_size()
11) Added load_default_teams().  It sets up the classic bronco teams into
a data structure called teaminfo.